LibAudio: Make WavWriter::write_samples take ReadonlyBytes
WavWriter::write_samples now takes the samples to be written as a
Span<u8 const>/ReadonlyBytes instead of as pointer and size. This might
not be easier to use, but it is less error-prone and avoids any
potential weird memory issues.

WavWriter now also uses the new Core::Stream::File instead of the old
Core::File. Simply switching to the new File did not quite work as the
file was stored as a RefPtr. It is now instead stored as an Optional. I
have no idea if this is correct, but it made the compiler happy. :^)

Next up is making the write functions return ErrorOr<size_t>.